Created conrad-blake_alternate.py
I was able to coordinate with blake  (thank you sir) using some of his base code to patch some of the missing logic. I then took the functions y'all wrote and mixed it in with what blake and I came up with. I think it all works now. I am submitting a pull request to be reviewed. The new code will be in a separate file called "conrad-blake_alternate".

Co-Authored-By: BlakeTheSnake34 <167281336+BlakeTheSnake34@users.noreply.github.com>